+++ resources/js/components/Footer.vue
@@ -7,7 +7,10 @@
                 </div>
             </div>
         </div>
-        <div class="content has-text-centered">
+        <div v-if="$route.name === 'settings'" class="content has-text-centered is-size-6">
+            <a class="has-text-grey" href="https://github.com/Bubka/2FAuth"><b>2FAuth</b> <font-awesome-icon :icon="['fab', 'github-alt']" /></a> - v{{ appVersion }}
+        </div>
+        <div v-else class="content has-text-centered">
             <router-link :to="{ name: 'settings' }" class="has-text-grey">{{ $t('settings.settings') }}</router-link> - <a class="has-text-grey" @click="logout">{{ $t('auth.sign_out') }}</a>
         </div>
     </footer>
@@ -27,21 +30,12 @@
         },
 
         methods: {
-
-            async logout(evt) {
+            logout() {
                 if(confirm(this.$t('auth.confirm.logout'))) {
 
-                    await this.axios.get('api/logout')
-
-                    localStorage.removeItem('jwt')
-                    localStorage.removeItem('user')
-
-                    delete this.axios.defaults.headers.common['Authorization']
-
-                    this.$router.push({ name: 'login' })
+                    this.appLogout()
                 }
-            },
-
+            }
         }
     };
 </script>