+++ resources/js/components/Modal.vue
@@ -1,27 +1,34 @@
 <template>
     <div class="modal modal-otp" v-bind:class="{ 'is-active': isActive }">
-      <div class="modal-background" @click.stop="closeModal"></div>
-      <div class="modal-content">
-        <section class="section">
-            <div class="columns is-centered">
-                <div class="column is-three-quarters">
-                    <div class="box has-text-centered has-background-black-ter is-shadowless">
-                        <slot></slot>
+        <div class="modal-background" @click.stop="closeModal"></div>
+        <div class="modal-content">
+            <section class="section">
+                <div class="columns is-centered">
+                    <div class="column is-three-quarters">
+                        <div class="box has-text-centered has-background-black-ter is-shadowless">
+                            <slot></slot>
+                        </div>
                     </div>
                 </div>
-            </div>
-        </section>
-      </div>
-      <button class="modal-close is-large" aria-label="close" @click.stop="closeModal"></button>
+            </section>
+        </div>
+        <div class="fullscreen-footer">
+            <!-- Close button -->
+            <label class="button is-dark is-rounded" @click.stop="closeModal">
+                {{ $t('commons.close') }}
+            </label>
+        </div>
     </div>
 </template>
 
 <script>
 export default {
     name: 'Modal',
+
     props: {
         value: Boolean,
     },
+
     computed: {
         isActive: {
             get () {
@@ -32,10 +39,12 @@ export default {
             }
         }
     },
+
     methods: {
         closeModal: function(event) {
             if (event) {
                 this.isActive = false
+                this.$notify({ clean: true })
                 this.$parent.$emit('modalClose')
             }
         }