+++ resources/js/components/index.js
@@ -1,25 +1,29 @@
-import Vue from 'vue'
-import Button from './Button'
-import FieldError from './FieldError'
-import FormWrapper from './FormWrapper'
-import FormField from './FormField'
-import FormSelect from './FormSelect'
-import FormSwitch from './FormSwitch'
-import FormButtons from './FormButtons'
-import Notification from './Notification'
-import VueFooter from './Footer'
+import Vue          from 'vue'
+import App          from './App'
+import Button       from './Button'
+import FieldError   from './FieldError'
+import FormWrapper  from './FormWrapper'
+import FormField    from './FormField'
+import FormSelect   from './FormSelect'
+import FormSwitch   from './FormSwitch'
+import FormCheckbox from './FormCheckbox'
+import FormButtons  from './FormButtons'
+import VueFooter    from './Footer'
+import Kicker       from './Kicker'
 
 // Components that are registered globaly.
 [
+    App,
 	Button,
     FieldError,
     FormWrapper,
     FormField,
     FormSelect,
     FormSwitch,
+    FormCheckbox,
     FormButtons,
-    Notification,
     VueFooter,
+    Kicker
 ].forEach(Component => {
 	Vue.component(Component.name, Component)
 })