+++ resources/lang/fr/twofaccounts.php
@@ -57,6 +57,7 @@ return [
         'not_readable' => 'Le scanner ne se charge pas. La caméra est-elle déjà utilisée ?',
         'no_cam_on_device' => 'Votre équipement ne dispose pas de caméra',
         'secured_context_required' => 'Contexte sécurisé requis (HTTPS ou localhost)',
+        'https_required' => 'HTTPS requis pour utiliser la caméra',
         'camera_not_suitable' => 'Votre équipement ne dispose pas d\'une caméra adaptée',
         'stream_api_not_supported' => 'L\'API Stream n\'est pas supportée par votre navigateur'
     ],